fix: get correct effective day and period start and end day while calculating prorated leaves
(cherry picked from commit d8dc6012c9c84416913305b89c05ba1870745bf6)
