+++ hrms/hr/utils.py
@@ -10,6 +10,7 @@ from frappe.query_builder import Criterion
 from frappe.query_builder.custom import ConstantColumn
 from frappe.utils import (
 	add_days,
+	add_months,
 	comma_and,
 	cstr,
 	flt,
@@ -20,6 +21,10 @@ from frappe.utils import (
 	get_last_day,
 	get_link_to_form,
 	get_number_format_info,
+	get_quarter_ending,
+	get_quarter_start,
+	get_year_ending,
+	get_year_start,
 	getdate,
 	nowdate,
 )
@@ -433,8 +438,12 @@ def get_monthly_earned_leave(
 		if pro_rated:
 			if not (period_start_date or period_end_date):
 				today_date = frappe.flags.current_date or getdate()
-				period_end_date = get_last_day(today_date)
-				period_start_date = get_first_day(today_date)
+
+				period_start_date, period_end_date = {
+					"Monthly": (get_first_day(today_date), get_last_day(today_date)),
+					"Quarterly": (get_quarter_start(today_date), get_quarter_ending(today_date)),
+					"Half-Yearly": (get_semester_start(today_date), get_semester_end(today_date)),
+				}.get(frequency)
 
 			earned_leaves = calculate_pro_rated_leaves(
 				earned_leaves, date_of_joining, period_start_date, period_end_date, is_earned_leave=True
@@ -508,27 +517,24 @@ def create_additional_leave_ledger_entry(allocation, leaves, date):
 
 
 def check_effective_date(from_date, today, frequency, allocate_on_day):
-	from dateutil import relativedelta
-
 	from_date = get_datetime(from_date)
 	today = frappe.flags.current_date or get_datetime(today)
-	rd = relativedelta.relativedelta(today, from_date)
 
 	expected_date = {
-		"First Day": get_first_day(today),
-		"Last Day": get_last_day(today),
-		"Date of Joining": from_date,
-	}[allocate_on_day]
+		"Monthly": {
+			"First Day": get_first_day(today),
+			"Last Day": get_last_day(today),
+			"Date of Joining": from_date,
+		},
+		"Quarterly": {
+			"First Day": get_quarter_start(today),
+			"Last Day": get_quarter_ending(today),
+		},
+		"Half-Yearly": {"First Day": get_semester_start(today), "Last Day": get_semester_end(today)},
+	}[frequency][allocate_on_day]
 
 	if expected_date.day == today.day:
-		if frequency == "Monthly":
-			return True
-		elif frequency == "Quarterly" and not rd.months % 3:
-			return True
-		elif frequency == "Half-Yearly" and not rd.months % 6:
-			return True
-		elif frequency == "Yearly" and not rd.months % 12:
-			return True
+		return True
 
 	return False
 
@@ -938,3 +944,17 @@ def get_exact_month_diff(string_ed_date: DateTimeLikeObject, string_st_date: Dat
 	if ed_date.day >= st_date.day:
 		diff += 1
 	return diff
+
+
+def get_semester_start(date):
+	if date.month <= 6:
+		return get_year_start(date)
+	else:
+		return add_months(get_year_start(date), 6)
+
+
+def get_semester_end(date):
+	if date.month > 6:
+		return get_year_ending(date)
+	else:
+		return add_months(get_year_ending(date), -6)
